backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "mcp>=1.13.1",
    "strands-agents>=1.7.1",
    "strands-agents-tools>=0.2.6",
    "fastapi>=0.115.6",
    "uvicorn[standard]>=0.34.0",
    "websockets>=14.1",
    "pydantic>=2.10.4",
    "pydantic-settings>=2.7.0",
    "python-multipart>=0.0.20",
    "aiofiles>=24.1.0",
    "python-dotenv>=1.0.1",
    "psutil>=6.1.1"
]
